Moving body, footer and header to common, then modifying styles accordingly
