+++ infl-styles/_button.scss
@@ -1,3 +1,6 @@
+@import '_theme';
+@import 'bourbon';
+
 $text-button-color : $medium-grey;
 $text-button-color-hover : lighten($text-button-color, 10%);
